+++ src/frontend/apps/desk/src/features/mail-domains/access-management/api/index.ts
@@ -1,4 +1,5 @@
 export * from './useMailDomainAccesses';
+export * from './useInvitationMailDomainAccesses';
 export * from './useUpdateMailDomainAccess';
 export * from './useCreateMailDomainAccess';
 export * from './useDeleteMailDomainAccess';