+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- ✨(front) add show invitations mails domains access #1040
+
 ## [1.22.2] - 2026-01-26
 
 ### Fixed